feat(abi): 拓宽 Pid 为 u64
u32 宽度源于未落地的 uni_tid = ((pid << 32) | tid) 拼装方案，属纯债务；
单调不复用的 PID 与 Zircon KOID 同理应取 64 位。三侧同步：
- shared：Pid = u64，ProcessId 统一为 proc::Pid；StartupBlockHeader
  pid/parent_pid 各 u64（头 40→48 字节）；ProcessCreateResult reserved
  提升为 u64，消除隐式 padding 以维持无 padding 写回前提
- kernel：Mailbox sender_pid 去转换
- rinlib：env 启动身份原子量改 AtomicU64；libprocess Spawned.pid 同步

+++ os/kernel/src/task/mailbox.rs
@@ -479,7 +479,7 @@ pub fn send(
             return Err(SystemCallError::WrongObjectType);
         }
         let message_header = MessageHeader::new(
-            thread.process.pid as u64,
+            thread.process.pid,
             badge,
             header.kind,
             header.payload_len,
